requirements_streamlit.txt
# Core trading system requirements
langchain==0.2.17
langgraph==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.19
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.7.0

# Data and analysis
yfinance>=0.2.43
finnhub-python>=2.4.20
pandas>=2.2.2
stockstats>=0.6.5
tavily-python>=0.3.5

# Streamlit UI requirements
streamlit>=1.28.0
plotly>=5.15.0
numpy>=1.24.0


